refactored `FeedItemsMapper` to the same structure as `ImageCommentsMapper`
