--- EssentialFeed/Feed API/RemoteFeedItem.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-struct RemoteFeedItem: Decodable {
-    var id: UUID
-    var description: String?
-    var location: String?
-    var image: URL
-}
